% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ad_pca}
\alias{ad_pca}
\alias{ad_pca.default}
\alias{ad_pca.data.frame}
\alias{ad_pca.matrix}
\alias{ad_pca.formula}
\alias{ad_pca.recipe}
\title{Fit a \code{ad_pca}}
\usage{
ad_pca(x, ...)

\method{ad_pca}{default}(x, ...)

\method{ad_pca}{data.frame}(x, y, ...)

\method{ad_pca}{matrix}(x, y, ...)

\method{ad_pca}{formula}(formula, data, ...)

\method{ad_pca}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{ad_pca} object.
}
\description{
\code{ad_pca()} fits a model.
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
mod <- ad_pca(predictors, outcome)

# Formula interface
mod2 <- ad_pca(mpg ~ ., mtcars)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- ad_pca(rec, mtcars)

}
