% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ames_new}
\alias{ames_new}
\title{Recent Ames Iowa Houses}
\source{
De Cock, D. (2011). "Ames, Iowa: Alternative to the Boston Housing
Data as an End of Semester Regression Project," \emph{Journal of Statistics
Education},  Volume 19, Number 3.

\url{https://www.cityofames.org/government/departments-divisions-a-h/city-assessor}

\url{http://jse.amstat.org/v19n3/decock/DataDocumentation.txt}

\url{http://jse.amstat.org/v19n3/decock.pdf}
}
\value{
\item{ames_new}{a tibble}
}
\description{
More data related to the set described by De Cock (2011) where data where
data were recorded for 2,930 properties in Ames IA.
}
\details{
This data sets includes three more properties added since the original
reference. There are less fields in this data set; only those that could be
transcribed from the assessor's office were included.
}
\keyword{datasets}
