% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoa-fit.R
\name{apd_aoa}
\alias{apd_aoa}
\alias{apd_aoa.default}
\alias{apd_aoa.data.frame}
\alias{apd_aoa.matrix}
\alias{apd_aoa.formula}
\alias{apd_aoa.recipe}
\title{Fit a \code{apd_aoa}}
\usage{
apd_aoa(x, ...)

\method{apd_aoa}{default}(x, ...)

\method{apd_aoa}{data.frame}(x, y = NULL, importance, ...)

\method{apd_aoa}{matrix}(x, y = NULL, importance, ...)

\method{apd_aoa}{formula}(formula, data, validation = NULL, importance, ...)

\method{apd_aoa}{recipe}(x, data, validation = NULL, importance, ...)
}
\arguments{
\item{x, y}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}

If \code{y} is \code{NULL}, then this function will use within-sample distances.
This may result in too large an area of applicability being calculated.}

\item{...}{Not currently used, but required for extensibility.}

\item{importance}{A data.frame with two columns: \code{Variable}, containing
the names of each variable in the training and validation data, and
\code{Importance}, containing the (raw or scaled) feature importance for each
variable.}

\item{formula}{A formula specifying the predictor terms on the right-hand
side.}

\item{data, validation}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing the predictors.
}

If \code{validation} is \code{NULL}, then this function will use within-sample distances.
This may result in too large an area of applicability being calculated.}
}
\value{
A \code{apd_aoa} object.
}
\description{
\code{apd_aoa()} fits a model.
}
\details{
This function calculates the "area of applicability" of a model, as
introduced by Meyer and Pebesma (2021). While the initial paper introducing
this method focused on spatial models, there is nothing inherently spatial
about the method; it can be used with any type of data.

The \code{importance} argument is structured to work with objects returned by the
vip package, using functions such as \link[vip:vi_permute]{vip::vi_permute}.
}
\examples{
\dontshow{if (rlang::is_installed("vip")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(vip)
train <- gen_friedman(1000, seed = 101)  # ?vip::gen_friedman
test <- train[701:1000, ]
train <- train[1:700, ]
pp <- stats::ppr(y ~ ., data = train, nterms = 11)
importance <- vi_permute(
  pp,
  target = "y",
  metric = "rsquared",
  pred_wrapper = predict
)

apd_aoa(y ~ ., train, test, importance = importance)
\dontshow{\}) # examplesIf}
}
\references{
H. Meyer and E. Pebesma. 2021. "Predicting into unknown space? Estimating
the area of applicability of spatial prediction models," Methods in Ecology
and Evolution 12(9), pp 1620 - 1633, doi: 10.1111/2041-210X.13650.
}
