% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolation-fit.R
\name{apd_isolation}
\alias{apd_isolation}
\alias{apd_isolation.default}
\alias{apd_isolation.data.frame}
\alias{apd_isolation.matrix}
\alias{apd_isolation.formula}
\alias{apd_isolation.recipe}
\title{Fit an isolation forest to estimate an applicability domain.}
\usage{
apd_isolation(x, ...)

\method{apd_isolation}{default}(x, ...)

\method{apd_isolation}{data.frame}(x, ...)

\method{apd_isolation}{matrix}(x, ...)

\method{apd_isolation}{formula}(formula, data, ...)

\method{apd_isolation}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Options to pass to \code{\link[isotree:isolation.forest]{isotree::isolation.forest()}}. Options should
not include \code{df}.}

\item{formula}{A formula specifying the predictor terms on the right-hand
side. No outcome should be specified.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing the predictors.
}}
}
\value{
A \code{apd_isolation} object.
}
\description{
\code{apd_isolation()} fits an isolation forest model.
}
\details{
background on isolation forests
}
\examples{
predictors <- mtcars[, -1]

# Data frame interface
mod <- apd_isolation(predictors)

# Formula interface
mod2 <- apd_isolation(mpg ~ ., mtcars)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- apd_isolation(rec, mtcars)
}
