% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicable-package.R
\docType{package}
\name{applicable-package}
\alias{applicable}
\alias{applicable-package}
\title{applicable: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Marly Cormar \email{marly@rstudio.com}

Authors:
\itemize{
  \item Max Kuhn \email{max@rstudio.com}
}

}
\keyword{internal}
