% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-score.R
\name{autoplot.apd_pca}
\alias{autoplot.apd_pca}
\title{Plot the distribution function for pcas}
\usage{
autoplot.apd_pca(object, ...)
}
\arguments{
\item{object}{An object produced by \code{apd_pca}.}

\item{...}{Not currently used.}
}
\value{
A \code{ggplot} object that shows the cdistribution function for each
principal component.
}
\description{
Plot the distribution function for pcas
}
