% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binary}
\alias{binary}
\alias{qsar_binary}
\alias{binary_tr}
\alias{binary_unk}
\title{Binary QSAR Data}
\value{
\item{binary_tr,binary_ukn}{data frame frames with 67 columns}
}
\description{
Binary QSAR Data
}
\details{
These data are from two different sources on quantitative
structure-activity relationship (QSAR) modeling and contain 67 predictors
that are either 0 or 1. The training set contains 4,330 samples and there
are five unknown samples (both from the \code{Mutagen} data in the \code{QSARdata}
package).
}
\examples{
data(qsar_binary)
str(binary_tr)
}
\keyword{datasets}
