% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{okc_binary}
\alias{okc_binary}
\alias{okc_binary_train}
\alias{okc_binary_test}
\title{OkCupid Binary Predictors}
\source{
Kim (2015), "OkCupid Data for Introductory Statistics and Data Science Courses", \emph{Journal of Statistics Education}, Volume 23, Number 2. \url{http://www.amstat.org/publications/jse/contents_2015.html}

Kuhn and Johnson (2020), \emph{Feature Engineering and Selection}, Chapman and Hall/CRC . \url{https://bookdown.org/max/FES/} and \url{https://github.com/topepo/FES}
}
\value{
\item{okc_binary_train. okc_binary_test}{data frame frames with 61 columns}
}
\description{
OkCupid Binary Predictors
}
\details{
Data originally from Kim (2015) includes a training and test set
consistent with Kuhn and Johnson (2020). Predictors include ethnicity
indicators and a set of keywords derived from text essay data.
}
\examples{
data(okc_binary)
str(okc_binary_train)
}
\keyword{datasets}
