% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.apd_hat_values}
\alias{print.apd_hat_values}
\title{Print number of predictors and principal components used.}
\usage{
\method{print}{apd_hat_values}(x, ...)
}
\arguments{
\item{x}{A \code{apd_hat_values} object.}

\item{...}{Not currently used, but required for extensibility.}
}
\value{
None
}
\description{
Print number of predictors and principal components used.
}
