% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.apd_pca}
\alias{print.apd_pca}
\title{Print number of predictors and principal components used.}
\usage{
\method{print}{apd_pca}(x, ...)
}
\arguments{
\item{x}{A \code{apd_pca} object.}

\item{...}{Not currently used, but required for extensibility.}
}
\value{
None
}
\description{
Print number of predictors and principal components used.
}
\examples{

model <- apd_pca(~ Sepal.Length + Sepal.Width, iris)
print(model)

}
