% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.apd_similarity}
\alias{print.apd_similarity}
\title{Print number of predictors and principal components used.}
\usage{
\method{print}{apd_similarity}(x, ...)
}
\arguments{
\item{x}{A \code{apd_similarity} object.}

\item{...}{Not currently used, but required for extensibility.}
}
\value{
None
}
\description{
Print number of predictors and principal components used.
}
