% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score.ad_pca}
\alias{score.ad_pca}
\title{Predict from a \code{ad_pca}}
\usage{
\method{score}{ad_pca}(object, new_data, type = "numeric", ...)
}
\arguments{
\item{object}{A \code{ad_pca} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{ad_pca}
}
\examples{
train <- mtcars[1:20,]
test <- mtcars[21:32, -1]

# Fit
mod <- ad_pca(mpg ~ cyl + log(drat), train)

# Predict, with preprocessing
score(mod, test)

}
