% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di-score.R
\name{score.apd_di}
\alias{score.apd_di}
\title{Predict from a \code{apd_di}}
\usage{
\method{score}{apd_di}(object, new_data, type = "numeric", ...)
}
\arguments{
\item{object}{A \code{apd_di} object.}

\item{new_data}{A data frame or matrix of new samples.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{apd_di}
}
\details{
The function computes the distance indices of the new data and
whether or not they are "inside" the area of applicability.
}
\examples{
\dontshow{if (rlang::is_installed("vip")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(vip)
train <- gen_friedman(1000, seed = 101)  # ?vip::gen_friedman
test <- train[701:1000, ]
train <- train[1:700, ]
pp <- stats::ppr(y ~ ., data = train, nterms = 11)
importance <- vi_permute(
  pp,
  target = "y",
  metric = "rsquared",
  pred_wrapper = predict
)

aoa <- apd_di(y ~ ., train, test, importance = importance)
score(aoa, test)
\dontshow{\}) # examplesIf}
}
