% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolation-score.R
\name{score.apd_isolation}
\alias{score.apd_isolation}
\title{Predict from a \code{apd_isolation}}
\usage{
\method{score}{apd_isolation}(object, new_data, type = "numeric", ...)
}
\arguments{
\item{object}{A \code{apd_isolation} object.}

\item{new_data}{A data frame or matrix of new samples.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{apd_isolation}
}
\details{
About the score
}
\examples{
train <- mtcars[1:20, ]
test <- mtcars[21:32, -1]

# Fit
mod <- apd_isolation(mpg ~ cyl + log(drat), train)

# Predict, with preprocessing
score(mod, test)
}
