% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-fit.R
\name{ad_pca}
\alias{ad_pca}
\alias{ad_pca.default}
\alias{ad_pca.data.frame}
\alias{ad_pca.matrix}
\alias{ad_pca.formula}
\alias{ad_pca.recipe}
\title{Fit a \code{ad_pca}}
\usage{
ad_pca(x, ...)

\method{ad_pca}{default}(x, ...)

\method{ad_pca}{data.frame}(x, ...)

\method{ad_pca}{matrix}(x, ...)

\method{ad_pca}{formula}(formula, data, ...)

\method{ad_pca}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{formula}{A formula specifying the predictor terms on the right-hand
side. No outcome should be specified.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing the predictors.
}}
}
\value{
A \code{ad_pca} object.
}
\description{
\code{ad_pca()} fits a model.
}
\examples{
predictors <- mtcars[, -1]

# Data frame interface
mod <- ad_pca(predictors)

# Formula interface
mod2 <- ad_pca(mpg ~ ., mtcars)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- ad_pca(rec, mtcars)

}
