% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicable-package.R
\docType{package}
\name{applicable-package}
\alias{applicable}
\alias{applicable-package}
\title{applicable: A Compilation of Applicability Domain Methods}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A modeling package compiling applicability domain methods in R. It combines different methods to measure the amount of extrapolation new samples can have from the training set. See Netzeva et al (2005) <doi:10.1177/026119290503300209> for an overview of applicability domains.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/applicable}
  \item \url{https://applicable.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/applicable/issues}
}

}
\author{
\strong{Maintainer}: Marly Gotti \email{marlygotti@gmail.com}

Authors:
\itemize{
  \item Max Kuhn \email{max@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
