% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicable-package.R
\docType{package}
\name{applicable-package}
\alias{applicable}
\alias{applicable-package}
\title{applicable: A Compilation of Applicability Domain Methods}
\description{
A modeling package compiling applicability domain methods in R.
    It combines different methods to measure the amount of extrapolation new
    samples can have from the training set.
}
\author{
\strong{Maintainer}: Marly Cormar \email{marlycormar@gmail.com}

Authors:
\itemize{
  \item Max Kuhn \email{max@rstudio.com}
}

}
\keyword{internal}
