% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.apd_pca}
\alias{autoplot.apd_pca}
\title{Plot the distribution function for pcas}
\usage{
autoplot.apd_pca(object, ...)
}
\arguments{
\item{object}{An object produced by \code{apd_pca}.}

\item{...}{An optional set of \code{dplyr} selectors, such as \code{dplyr::matches()} or
\code{dplyr::starts_with()} for selecting which variables should be shown in the
plot.}
}
\value{
A \code{ggplot} object that shows the cdistribution function for each
principal component.
}
\description{
Plot the distribution function for pcas
}
