% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binary}
\alias{binary}
\alias{qsar_binary}
\alias{binary_tr}
\alias{binary_unk}
\title{Binary QSAR Data}
\value{
\item{binary_tr,binary_ukn}{data frame frames with 49 columns}
}
\description{
Binary QSAR Data
}
\details{
These data are from two different sources on quantitative
structure-activity relationship (QSAR) modeling and contain 49 predictors
that are either 0 or 1. The training set contains 3796 samples (from the
\code{caco} data set in the \code{QSARdata} package) and there are 80 unknown samples
(taken from the \code{bbb2} data in the same package).

The training set was collected from a general set used to measure
permeability while the new samples are from a set of molecules used to study
brain penetration (presumably for neuroscience related purposes).
}
\examples{
data(qsar_binary)
str(binary_tr)
}
\keyword{datasets}
